location loeschen

+++ src/main/java/de/deftone/demo/repo/LocationRepo.java
@@ -3,5 +3,5 @@ package de.deftone.demo.repo;
 import de.deftone.demo.model.Location;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocationRepo extends JpaRepository<Location, Integer> {
+public interface LocationRepo extends JpaRepository<Location, Long> {
 }
